Time: 0 ms (100.00%), Space: 5.8 MB (88.13%) - LeetHub

+++ 62-unique-paths/62-unique-paths.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int uniquePaths(int m, int n) {
         int N = m + n - 2, r = m - 1;
-        double res = 1;
+        long long int res = 1;
         
         for (int i = 1; i <= r; i++)
             res = (res * (N - r + i)) / i;
